Separated declarations and implementations of UDTs. Provided appropriate includes.
