+++ BuildingWrapper.cpp
@@ -1,0 +1,8 @@
+#include "BuildingWrapper.h"
+
+BuildingWrapper::BuildingWrapper(Building* b) : pointerToBuilding(b) {}
+
+BuildingWrapper::~BuildingWrapper()
+{
+    delete pointerToBuilding;
+}
