+++ BuildingWrapper.h
@@ -1,0 +1,8 @@
+#include "Building.h"
+
+struct BuildingWrapper
+{
+    Building* pointerToBuilding = nullptr;
+    BuildingWrapper(Building* b);
+    ~BuildingWrapper();
+};
