+++ DrillWrapper.cpp
@@ -1,0 +1,9 @@
+#include "DrillWrapper.h"
+
+DrillWrapper::DrillWrapper(Drill* d) : pointerToDrill(d) {}
+
+DrillWrapper::~DrillWrapper()
+{
+    delete pointerToDrill;
+}
+
