+++ DrillWrapper.h
@@ -1,0 +1,8 @@
+#include "Drill.h"
+
+struct DrillWrapper
+{
+    Drill* pointerToDrill = nullptr;
+    DrillWrapper(Drill* d);
+    ~DrillWrapper();
+};
