+++ ElevatorWrapper.cpp
@@ -1,0 +1,9 @@
+#include "ElevatorWrapper.h"
+
+ElevatorWrapper::ElevatorWrapper(Elevator* ev) : pointerToEV(ev) {}
+
+ElevatorWrapper::~ElevatorWrapper()
+{
+    delete pointerToEV;
+}
+
