+++ ElevatorWrapper.h
@@ -1,0 +1,8 @@
+#include "Elevator.h"
+
+struct ElevatorWrapper
+{
+    Elevator* pointerToEV = nullptr;
+    ElevatorWrapper(Elevator* ev);
+    ~ElevatorWrapper();
+};
