+++ PencilCaseWrapper.cpp
@@ -1,0 +1,9 @@
+#include "PencilCaseWrapper.h"
+
+PencilCaseWrapper::PencilCaseWrapper(PencilCase* pc) : pointerToPC(pc) {}
+
+PencilCaseWrapper::~PencilCaseWrapper()
+{
+    delete pointerToPC;
+}
+
