+++ PencilCaseWrapper.h
@@ -1,0 +1,8 @@
+#include "PencilCase.h"
+
+struct PencilCaseWrapper
+{
+    PencilCase* pointerToPC = nullptr;
+    PencilCaseWrapper(PencilCase* pc);
+    ~PencilCaseWrapper();
+};
