+++ SchoolBagWrapper.cpp
@@ -1,0 +1,8 @@
+#include "SchoolBagWrapper.h"
+
+SchoolBagWrapper::SchoolBagWrapper(SchoolBag* sb) : pointerToSB(sb) {}
+
+SchoolBagWrapper::~SchoolBagWrapper()
+{
+    delete pointerToSB;
+}
