+++ SchoolBagWrapper.h
@@ -1,0 +1,8 @@
+#include "SchoolBag.h"
+
+struct SchoolBagWrapper
+{
+    SchoolBag* pointerToSB = nullptr;
+    SchoolBagWrapper(SchoolBag* sb);
+    ~SchoolBagWrapper();
+};
